Encrypt tokens
Now we don't use gogdl, so we can store tokens with safer way by encrypting like its done in for ex official Epic plugin

+++ src/GogOssLibrarySettingsView.xaml.cs
@@ -273,7 +273,14 @@ namespace GogOssLibraryNS
                 if (answer == MessageBoxResult.Yes)
                 {
                     view.DeleteDomainCookies(".gog.com");
-                    File.Delete(GogOss.TokensPath);
+                    if (File.Exists(GogOss.TokensPath))
+                    {
+                        File.Delete(GogOss.TokensPath);
+                    }
+                    if (File.Exists(GogOss.EncryptedTokensPath))
+                    {
+                        File.Delete(GogOss.EncryptedTokensPath);
+                    }
                     UpdateAuthStatus();
                 }
                 else
